volumes/app/requirements.txt
amqp==5.1.1
asgiref==3.6.0
Django==4.2
sqlparse==0.4.4
psycopg==3.1.8
gunicorn==20.1.0
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.2
django-money==3.1.0
celery==5.2.7